Remove `BeamResource.syncPropertiesToInternal`
This greatly simplifies the logic by always using `resolvedKeyValues()` method when fetching values either to resolve properties or to dump values to disk.

This avoids the issue where BeamReference cannot be property synced from a property to an internal value and back. When this happened the value was nulled out and lost.