Avoid calling syncInternalToProperties until providers are loaded
This moves the logic for `syncInternalToProperties` from `ResourceNode` to `BeamBaseResource`.

This prevents setting properties from failing during the initial parser pass to load providers. This happens because on the first parser pass resources are loaded as `ResourceNode` objects rather than their final implementation classes so the properties don't exists during this parser pass.