Refactor serialization to output maps/lists with references
If a map/list contained a reference to an object, for example a list of references to subnets and their ids, it was being saved to state as the final subnet-ids instead of as a reference. This was causing state to lose some dependency data, and as a result deletion was not always happening in the correct order.