Evaluate and create resources within a virtual resource
Right now this is not a complete solution since the resources created inside the virtual resource are not scoped correctly.

From inside the virtual resource the resources should be accessible by their base names (i.e. the name of the resource without any scoping, such as 'vpc').

From outside the virtual resource the resources should be accessible by their scoped names (i.e. the name of the virtual resource usages + the name of the resource, such as 'network.vpc').