Fixes configured fields to be set correctly.
This was caused by bodyScope to map conversion in NodeEvaluator.visitResource method that was necessary for extends directive to work during a workflow. Since a scope instance is necessary during Diffable.initialize for configured fields to be detected correctly, the fix was to detect this and handle it specially.

Eventually, we should refactor Scope.keyNodes and valueNodes so that this isn't necessary.