Copy state back to pending resources after re-evaluating them
Without this change there are situations where output values are lost during the diff triggering Gyro to try to make changes that are not needed. Further, these changes want are invalid. It tries to go from a valid value to "<output>". The "<output>" is a placeholder but in this scenario it's never updated before calling cloud APIs causing those APIs to fail.