Refactor out workflow related work
Most of the workflow related code will now live inside the plugin `gyro-workflow-plugin`.

Most of the business logic has just been moved to the new plugin.

Some of the existing interfaces/classes have intertwined deendency in the core, and the bare minimum code remains which serves mostly as an interaction layer for the workflow plugin.